+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - - -
 
+## Version `3.3.2`
+- Patch error in shard delete 'by bag' function
+
 ## Version `3.3.1`
 - Patch for 3.4.1 client compatibility